+++ application/core/src/main/java/com/victorprado/donus/core/usecase/performtransaction/UpdateBankAccountBalance.java
@@ -3,7 +3,9 @@ package com.victorprado.donus.core.usecase.performtransaction;
 import com.victorprado.donus.core.entity.BankAccount;
 import com.victorprado.donus.core.exception.DataProviderException;
 
+import java.math.BigDecimal;
+
 public interface UpdateBankAccountBalance {
 
-    void updateBalance(BankAccount bankAccount, Double newValue) throws DataProviderException;
+    void updateBalance(BankAccount bankAccount, BigDecimal newValue) throws DataProviderException;
 }
