+++ application/dataproviders/src/main/java/com/victorprado/donus/database/rowmapper/BankAccountRowMapper.java
@@ -3,6 +3,7 @@ package com.victorprado.donus.database.rowmapper;
 import com.victorprado.donus.core.entity.BankAccount;
 import org.springframework.jdbc.core.RowMapper;
 
+import java.math.BigDecimal;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.time.LocalDateTime;
@@ -20,7 +21,7 @@ public class BankAccountRowMapper implements RowMapper<BankAccount> {
     public BankAccount mapRow(ResultSet rs, int rowNum) throws SQLException {
         BankAccount bankAccount = new BankAccount();
         bankAccount.setId(rs.getString(ID));
-        bankAccount.setBalance(rs.getDouble(BALANCE));
+        bankAccount.setBalance(BigDecimal.valueOf(rs.getDouble(BALANCE)));
         bankAccount.setNumber(rs.getString(NUMBER));
 
         bankAccount.setCreatedDate(rs.getObject(CREATED_DATE, LocalDateTime.class));
