+++ application/entrypoints/src/main/java/rest/bankaccount/TransferTransactionDTO.java
@@ -1,9 +1,11 @@
 package rest.bankaccount;
 
+import java.math.BigDecimal;
+
 public class TransferTransactionDTO implements Dto {
 
     private String destinationAccountNumber;
-    private Double value;
+    private BigDecimal value;
     private String when;
 
     public String getDestinationAccountNumber() {
@@ -14,11 +16,11 @@ public class TransferTransactionDTO implements Dto {
         this.destinationAccountNumber = destinationAccountNumber;
     }
 
-    public Double getValue() {
+    public BigDecimal getValue() {
         return value;
     }
 
-    public void setValue(Double value) {
+    public void setValue(BigDecimal value) {
         this.value = value;
     }
